implement: 68/69 tests run under mocha with temporary "npm run test:new"
