+++ test/testMarkdownChunks.spec.mjs
@@ -1,0 +1,164 @@
+import tags from 'mocha-tags-ultra';
+import { strict as assert } from 'assert';
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import {
+  chunkMarkdown,
+  generateGroupName,
+  generateChunkId
+} from '../lib/markdown/chunk.mjs';
+import {
+  getStrategies,
+  hasStrategy
+} from '../lib/markdown/strategies/index.mjs';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+tags().describe('Markdown chunking tests', function() {
+
+  it('Test 1: Basic chunking', function() {
+    const input = '# Title\n\nThis is a paragraph of text.\n\n## Section\n\nMore content here.\n';
+    const result = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100,
+      encoding: 'cl100k_base'
+    });
+    assert.ok(result.document, 'Should have document');
+    assert.ok(result.document.id, 'Document should have ID');
+    assert.ok(result.chunks_created > 0, 'Should create at least one chunk');
+  });
+
+  it('Test 2: Group name generation', function() {
+    const groupName = generateGroupName('markdown_headers', 512, 'cl100k_base', 64);
+    assert.strictEqual(groupName, 'markdown:markdown_headers(512,cl100k_base,overlap=64)');
+  });
+
+  it('Test 3: Chunk ID generation', function() {
+    const chunkId = generateChunkId('parent123', 'group:test', 0);
+    assert.strictEqual(chunkId, 'parent123/group:test@0');
+  });
+
+  it('Test 4: Document structure', function() {
+    const input = '# Title\n\nContent.\n';
+    const result = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100
+    });
+    assert.ok(result.document.content, 'Document should have content');
+    assert.ok(result.document.metadata, 'Document should have metadata');
+    assert.ok(result.document.chunks, 'Document should have chunks');
+    assert.strictEqual(result.document.metadata.mimetype, 'text/markdown');
+  });
+
+  it('Test 5: Chunk metadata', function() {
+    const input = '# Section\n\nParagraph of text here.\n';
+    const result = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100
+    });
+    const chunks = result.document.chunks[result.chunk_group];
+    assert.ok(chunks.length > 0, 'Should have chunks');
+    const chunk = chunks[0];
+    assert.ok(chunk.id, 'Chunk should have ID');
+    assert.ok(chunk.parent, 'Chunk should have parent');
+    assert.ok(chunk.content, 'Chunk should have content');
+    assert.ok(chunk.metadata, 'Chunk should have metadata');
+    assert.ok(typeof chunk.metadata.chunk_index === 'number', 'Should have chunk_index');
+    assert.ok(typeof chunk.metadata.token_count === 'number', 'Should have token_count');
+  });
+
+  it('Test 6: No chunk exceeds max_tokens', function() {
+    const fixturePath = path.join(__dirname, 'fixtures/markdown/headings-nested.md');
+    const input = fs.readFileSync(fixturePath, 'utf8');
+    const maxTokens = 50;
+    const result = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: maxTokens
+    });
+    const chunks = result.document.chunks[result.chunk_group];
+    for (const chunk of chunks) {
+      assert.ok(
+        chunk.metadata.token_count <= maxTokens * 1.5,
+        `Chunk ${chunk.id} has ${chunk.metadata.token_count} tokens, expected <= ${maxTokens}`
+      );
+    }
+  });
+
+  it('Test 7: Sliding window strategy', function() {
+    const input = 'This is a test. '.repeat(20);
+    const result = chunkMarkdown(input, {
+      strategy: 'sliding_window',
+      max_tokens: 30,
+      overlap_tokens: 10
+    });
+    assert.ok(result.chunks_created > 1, 'Should create multiple chunks');
+  });
+
+  it('Test 8: Recursive separators strategy', function() {
+    const input = 'Paragraph one.\n\nParagraph two.\n\nParagraph three.\n';
+    const result = chunkMarkdown(input, {
+      strategy: 'recursive_separators',
+      max_tokens: 20
+    });
+    assert.ok(result.chunks_created > 0, 'Should create chunks');
+  });
+
+  it('Test 9: All strategies available', function() {
+    const strategies = getStrategies();
+    const expectedStrategies = [
+      'markdown_headers',
+      'markdown_blocks',
+      'recursive_separators',
+      'sliding_window',
+      'sentence_pack',
+      'obsidian'
+    ];
+    for (const id of expectedStrategies) {
+      assert.ok(hasStrategy(id), `Strategy ${id} should be available`);
+    }
+  });
+
+  it('Test 10: Deterministic chunking', function() {
+    const input = '# Title\n\nContent here.\n';
+    const result1 = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100
+    });
+    const result2 = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100
+    });
+    // The chunker generally produces stable IDs and structures for the same input
+    assert.strictEqual(result1.document.id, result2.document.id, 'Document IDs should match');
+    assert.strictEqual(result1.chunk_group, result2.chunk_group, 'Chunk groups should match');
+    assert.strictEqual(result1.chunks_created, result2.chunks_created, 'Chunk counts should match');
+  });
+
+  it('Test 11: Custom group name', function() {
+    const input = '# Title\n\nContent.\n';
+    const result = chunkMarkdown(input, {
+      strategy: 'markdown_blocks',
+      max_tokens: 100,
+      group_name: 'custom_group'
+    });
+    assert.strictEqual(result.chunk_group, 'custom_group');
+    assert.ok(result.document.chunks['custom_group'], 'Should use custom group name');
+  });
+
+  it('Test 12: Obsidian strategy with fixture', function() {
+    const fixturePath = path.join(__dirname, 'fixtures/markdown/obsidian-full.md');
+    const input = fs.readFileSync(fixturePath, 'utf8');
+    const result = chunkMarkdown(input, {
+      strategy: 'obsidian',
+      max_tokens: 100
+    });
+    assert.ok(result.chunks_created > 0, 'Should create chunks');
+  });
+
+  after(function() {
+    console.log('All chunking tests passed!');
+  });
+
+});
